add interface interaction to task screen

+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/screen/tasks/TasksScreen.kt
@@ -34,33 +34,26 @@ import androidx.compose.ui.draw.rotate
 import androidx.compose.ui.platform.LocalLayoutDirection
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.LayoutDirection
 import androidx.compose.ui.unit.dp
 import com.sanaa.tudee_assistant.R
 import com.sanaa.tudee_assistant.presentation.composable.CustomDatePickerDialog
-import com.sanaa.tudee_assistant.presentation.composable.bottomSheet.DeleteComponent
 import com.sanaa.tudee_assistant.presentation.composable.TaskStatusTabs
+import com.sanaa.tudee_assistant.presentation.composable.bottomSheet.DeleteComponent
+import com.sanaa.tudee_assistant.presentation.composable.bottomSheet.task.AddEditTaskScreen
+import com.sanaa.tudee_assistant.presentation.composable.bottomSheet.task.TaskDetailsComponent
 import com.sanaa.tudee_assistant.presentation.designSystem.component.DayItem
 import com.sanaa.tudee_assistant.presentation.designSystem.component.TudeeSnackBar
 import com.sanaa.tudee_assistant.presentation.designSystem.component.button.FloatingActionButton
 import com.sanaa.tudee_assistant.presentation.designSystem.theme.Theme
-import com.sanaa.tudee_assistant.presentation.model.TaskUiStatus
 import com.sanaa.tudee_assistant.presentation.route.TasksScreenRoute
-import com.sanaa.tudee_assistant.presentation.composable.bottomSheet.task.AddEditTaskScreen
-import com.sanaa.tudee_assistant.presentation.composable.bottomSheet.task.TaskDetailsComponent
 import com.sanaa.tudee_assistant.presentation.state.TaskUiState
-import com.sanaa.tudee_assistant.presentation.utils.DataProvider
 import com.sanaa.tudee_assistant.presentation.utils.DateFormater
 import com.sanaa.tudee_assistant.presentation.utils.DateFormater.getShortMonthName
 import kotlinx.coroutines.launch
-import kotlinx.datetime.Clock
 import kotlinx.datetime.DateTimeUnit
-import kotlinx.datetime.LocalDate
-import kotlinx.datetime.TimeZone
 import kotlinx.datetime.minus
 import kotlinx.datetime.plus
-import kotlinx.datetime.toLocalDateTime
 import org.koin.androidx.compose.koinViewModel
 
 
@@ -78,17 +71,7 @@ fun TasksScreen(
 
     TasksScreenContent(
         state = state,
-        onDateSelected = { date -> viewModel.onDueDateChange(date) },
-        onTaskSwipe = { task ->
-            viewModel.onTaskSwipeToDelete(task)
-            false
-        },
-        onTaskClick = viewModel::onTaskClick,
-        onDismissTaskViewDetails = { viewModel.onShowTaskDetailsDialogChange(false) },
-        onMoveToTaskViewDetails = viewModel::onMoveTaskToAnotherStatus,
-        onDeleteClick = viewModel::onTaskDeleted,
-        onDeleteDismiss = viewModel::onTaskDeletedDismiss,
-        onShowSnackBar = viewModel::onSnackBarShown,
+        interactionListener = viewModel,
         modifier = modifier,
     )
 }
@@ -97,14 +80,7 @@ fun TasksScreen(
 @Composable
 fun TasksScreenContent(
     state: TasksScreenUiState,
-    onTaskSwipe: (TaskUiState) -> Boolean,
-    onDateSelected: (LocalDate) -> Unit,
-    onTaskClick: (TaskUiState) -> Unit,
-    onDismissTaskViewDetails: () -> Unit,
-    onMoveToTaskViewDetails: () -> Unit,
-    onDeleteClick: () -> Unit,
-    onDeleteDismiss: () -> Unit,
-    onShowSnackBar: () -> Unit,
+    interactionListener: TaskInteractionListener,
     modifier: Modifier = Modifier,
 ) {
 
@@ -114,18 +90,15 @@ fun TasksScreenContent(
     var taskToEdit by remember { mutableStateOf<TaskUiState?>(null) }
     var showDialog by remember { mutableStateOf(false) }
     var showAddTaskBottomSheet by remember { mutableStateOf(false) }
-    var daysInMonth by
-    remember {
+    var daysInMonth by remember {
         mutableStateOf(
             DateFormater.getLocalDatesInMonth(
-                state.selectedDate.year,
-                state.selectedDate.monthNumber
+                state.selectedDate.year, state.selectedDate.monthNumber
             )
         )
     }
     Box(
-        modifier = modifier
-            .fillMaxSize()
+        modifier = modifier.fillMaxSize()
     ) {
         Column(
             modifier = modifier
@@ -152,14 +125,11 @@ fun TasksScreenContent(
                         .padding(6.dp)
                         .clickable {
                             val nextMonth = state.selectedDate.minus(1, DateTimeUnit.MONTH)
-                            onDateSelected(nextMonth)
-                            daysInMonth =
-                                (DateFormater.getLocalDatesInMonth(
-                                    nextMonth.year,
-                                    nextMonth.monthNumber
-                                ))
-                        },
-                    contentAlignment = Alignment.Center
+                            interactionListener.onDateSelected(nextMonth)
+                            daysInMonth = (DateFormater.getLocalDatesInMonth(
+                                nextMonth.year, nextMonth.monthNumber
+                            ))
+                        }, contentAlignment = Alignment.Center
                 ) {
                     Icon(
                         painter = painterResource(R.drawable.icon_left_arrow),
@@ -187,8 +157,7 @@ fun TasksScreenContent(
                             .rotate(if (LocalLayoutDirection.current == LayoutDirection.Rtl) 90f else -90f)
                             .clickable {
                                 showDialog = true
-                            }
-                    )
+                            })
                 }
 
                 Box(
@@ -198,14 +167,11 @@ fun TasksScreenContent(
                         .padding(6.dp)
                         .clickable {
                             val previousMonth = state.selectedDate.plus(1, DateTimeUnit.MONTH)
-                            onDateSelected(previousMonth)
-                            daysInMonth =
-                                (DateFormater.getLocalDatesInMonth(
-                                    previousMonth.year,
-                                    previousMonth.monthNumber
-                                ))
-                        },
-                    contentAlignment = Alignment.Center
+                            interactionListener.onDateSelected(previousMonth)
+                            daysInMonth = (DateFormater.getLocalDatesInMonth(
+                                previousMonth.year, previousMonth.monthNumber
+                            ))
+                        }, contentAlignment = Alignment.Center
                 ) {
                     Icon(
                         painter = painterResource(R.drawable.icon_left_arrow),
@@ -236,50 +202,45 @@ fun TasksScreenContent(
             ) {
                 itemsIndexed(daysInMonth) { index, date ->
                     DayItem(
-                        dayDate = date,
-                        isSelected = date == state.selectedDate,
-                        onClick = {
-                            onDateSelected(date)
-                        }
-                    )
+                        dayDate = date, isSelected = date == state.selectedDate, onClick = {
+                            interactionListener.onDateSelected(date)
+                        })
                 }
             }
 
             if (showDialog) {
                 CustomDatePickerDialog(
                     onDateSelected = { selectedDateMillis: Long? ->
-                        selectedDateMillis?.let {
-                            val date = DateFormater.formatLongToDate(selectedDateMillis)
-                            onDateSelected(date)
-                            daysInMonth =
-                                (DateFormater.getLocalDatesInMonth(date.year, date.monthNumber))
-                        }
-                    },
-                    onDismiss = { showDialog = false },
-                    initialSelectedDate = state.selectedDate
+                    selectedDateMillis?.let {
+                        val date = DateFormater.formatLongToDate(selectedDateMillis)
+                        interactionListener.onDateSelected(date)
+                        daysInMonth =
+                            (DateFormater.getLocalDatesInMonth(date.year, date.monthNumber))
+                    }
+                }, onDismiss = { showDialog = false }, initialSelectedDate = state.selectedDate
                 )
             }
-            TaskStatusTabs(state, onTaskSwipe, onTaskClick)
-            if (state.selectedTask != null && state.showTaskDetailsBottomSheet)
-                TaskDetailsComponent(
-                    task = state.selectedTask,
-                    onDismiss = onDismissTaskViewDetails,
-                    onEditClick = { task ->
-                        onDismissTaskViewDetails()
-                        taskToEdit = task
-                        showEditTaskBottomSheet = true
-                    },
-                    onMoveToClicked = onMoveToTaskViewDetails,
-                    modifier = Modifier.padding(horizontal = 16.dp)
-                )
-            if (state.selectedTask != null && state.showDeleteTaskBottomSheet)
-                DeleteComponent(
-                    onDismiss = onDeleteDismiss,
-                    onDeleteClicked = {
-                        onDeleteClick()
-                    },
-                    title = stringResource(R.string.delete_task_title),
-                )
+            TaskStatusTabs(
+                state,
+                interactionListener::onTaskSwipeToDelete,
+                interactionListener::onTaskClicked
+            )
+            if (state.selectedTask != null && state.showTaskDetailsBottomSheet) TaskDetailsComponent(
+                task = state.selectedTask,
+                onDismiss = { interactionListener.onDismissTaskDetails(false) },
+                onEditClick = { task ->
+                    interactionListener.onDismissTaskDetails(false)
+                    taskToEdit = task
+                    showEditTaskBottomSheet = true
+                },
+                onMoveToClicked = interactionListener::onMoveTaskToAnotherStatus,
+                modifier = Modifier.padding(horizontal = 16.dp)
+            )
+            if (state.selectedTask != null && state.showDeleteTaskBottomSheet) DeleteComponent(
+                onDismiss = interactionListener::onDeleteDismiss,
+                onDeleteClicked = interactionListener::onDeleteTask,
+                title = stringResource(R.string.delete_task_title),
+            )
             if (showAddTaskBottomSheet) {
                 AddEditTaskScreen(
                     isEditMode = false,
@@ -289,49 +250,38 @@ fun TasksScreenContent(
                         showAddTaskBottomSheet = false
                         coroutineScope.launch {
                             snackBarHostState.showSnackbar(
-                                message = "Task added successfully",
-                                withDismissAction = true
+                                message = "Task added successfully", withDismissAction = true
                             )
                         }
                     },
                     onError = { errorMessage ->
                         coroutineScope.launch {
                             snackBarHostState.showSnackbar(
-                                message = errorMessage,
-                                withDismissAction = true
+                                message = errorMessage, withDismissAction = true
                             )
                         }
-                    }
-                )
+                    })
             }
 
             if (showEditTaskBottomSheet && taskToEdit != null) {
-                AddEditTaskScreen(
-                    isEditMode = true,
-                    taskToEdit = taskToEdit,
-                    onDismiss = {
-                        showEditTaskBottomSheet = false
-                        taskToEdit = null
-                    },
-                    onSuccess = {
-                        showEditTaskBottomSheet = false
-                        taskToEdit = null
-                        coroutineScope.launch {
-                            snackBarHostState.showSnackbar(
-                                message = "Task updated successfully",
-                                withDismissAction = true
-                            )
-                        }
-                    },
-                    onError = { errorMessage ->
-                        coroutineScope.launch {
-                            snackBarHostState.showSnackbar(
-                                message = errorMessage,
-                                withDismissAction = true
-                            )
-                        }
+                AddEditTaskScreen(isEditMode = true, taskToEdit = taskToEdit, onDismiss = {
+                    showEditTaskBottomSheet = false
+                    taskToEdit = null
+                }, onSuccess = {
+                    showEditTaskBottomSheet = false
+                    taskToEdit = null
+                    coroutineScope.launch {
+                        snackBarHostState.showSnackbar(
+                            message = "Task updated successfully", withDismissAction = true
+                        )
                     }
-                )
+                }, onError = { errorMessage ->
+                    coroutineScope.launch {
+                        snackBarHostState.showSnackbar(
+                            message = errorMessage, withDismissAction = true
+                        )
+                    }
+                })
             }
         }
 
@@ -349,7 +299,7 @@ fun TasksScreenContent(
             iconRes = R.drawable.note_add
         )
 
-        var successMessageText : String? = null
+        var successMessageText: String? = null
         state.successMessageStringId?.let {
             successMessageText = stringResource(state.successMessageStringId)
         }
@@ -357,66 +307,22 @@ fun TasksScreenContent(
         LaunchedEffect(state.successMessageStringId) {
             state.successMessageStringId?.let {
                 snackBarHostState.showSnackbar(successMessageText!!)
-                onShowSnackBar()
+                interactionListener.onShowSnackbar()
             }
         }
-        var errorMessageText : String? = null
+        var errorMessageText: String? = null
         state.errorMessageStringId?.let {
             successMessageText = stringResource(state.errorMessageStringId)
         }
         LaunchedEffect(state.successMessageStringId) {
             state.errorMessageStringId?.let {
                 snackBarHostState.showSnackbar(errorMessageText!!)
-                onShowSnackBar()
+                interactionListener.onShowSnackbar()
             }
         }
 
         TudeeSnackBar(
-            snackBarHostState = snackBarHostState,
-            isError = state.errorMessageStringId != null
+            snackBarHostState = snackBarHostState, isError = state.errorMessageStringId != null
         )
     }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun TasksScreenPreview() {
-
-    var tasksState by remember {
-        mutableStateOf(
-            TasksScreenUiState(
-                currentDateTasks = DataProvider.getTasksSample(),
-                errorMessageStringId = null,
-                successMessageStringId = null,
-                showAddTaskBottomSheet = false,
-                showEditTaskBottomSheet = false,
-                showTaskDetailsBottomSheet = false,
-                showDeleteTaskBottomSheet = false,
-                selectedStatusTab = TaskUiStatus.DONE,
-                selectedTask = null,
-                selectedDate = Clock.System.now().toLocalDateTime(TimeZone.UTC).date,
-            )
-        )
-    }
-
-    TasksScreenContent(
-        tasksState,
-        onDateSelected = { date ->
-            tasksState = tasksState.copy(selectedDate = date)
-            tasksState = tasksState.copy(currentDateTasks = DataProvider.getTasksSample())
-        },
-        onTaskSwipe = { task ->
-            tasksState =
-                tasksState.copy(currentDateTasks = tasksState.currentDateTasks.filterNot { item -> item == task })
-            false
-        },
-        onTaskClick = { task ->
-            tasksState = tasksState.copy(selectedTask = task, showTaskDetailsBottomSheet = true)
-        },
-        onDismissTaskViewDetails = { tasksState = tasksState.copy(showTaskDetailsBottomSheet = false) },
-        onMoveToTaskViewDetails = {},
-        onDeleteClick = {},
-        onDeleteDismiss = {},
-        onShowSnackBar = {},
-    )
 }
